# optimize convertIstioListenerToWrapper performance