@istio-testing 2023-07-17
@wulianglongrd: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
unit-tests-arm64_istio | 103fc9b9c81b44e6a9d004b3d1b2fbc9af1c8ac9 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46043/unit-tests-arm64_istio/1680946730943647744) | true | `/test unit-tests-arm64`
lint_istio | 103fc9b9c81b44e6a9d004b3d1b2fbc9af1c8ac9 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46043/lint_istio/1680946741769146368) | true | `/test lint`
unit-tests_istio | 103fc9b9c81b44e6a9d004b3d1b2fbc9af1c8ac9 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46043/unit-tests_istio/1680946730897510400) | true | `/test unit-tests`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->