@hzxuzhonghu 2023-07-18
Is the optimization mostly based on host match? If so i would ask how many hosts do we usually set on sidecar? 

I doubt the complexity introduced weigh over the perf improve