@wulianglongrd 2023-07-18
> Is the optimization mostly based on host match? If so i would ask how many hosts do we usually set on sidecar?

Yes, a Sidecar CR's egress usually contains 100~400 hosts. We are currently not using wildcard host.

> I doubt the complexity introduced weigh over the perf improve

I'm also a little concern. This PR currently has a bug (vs.host is a wildcard scenario not considered). I'll take another look.