@hzxuzhonghu 2023-07-18
>Yes, a Sidecar CR's egress usually contains 100~400 hosts. We are currently not using wildcard host.

Confusing to me, what is your application do while needing to access more than 100 service