@wulianglongrd 2023-07-28
forgot there was a discussion here, so sorry.

> how many entries do you have here in real world?

hello @howardjohn , about 50. It is not only related to the number of SEs, but also related to the number of `SE.hosts`.


> Confusing to me, what is your application do while needing to access more than 100 service

hello @hzxuzhonghu , there are not so many real services, 5~20 are more common. This is mainly caused by the large size of `SE.hosts`. Each service currently using mesh has about 15 clusters (some clusters are deployed independently for specific usage scenarios), each with a unique host. Currently, our self-developed tools for tracking service dependencies in real time are not yet mature(use it to update `Sidecar.egress.hosts` in real time). Therefore, if a consumer depends on a provider, we will write all clusters (hosts) of the provider to `Sidecar.egress.hosts`. There are other reasons, like we have multiple namespaces for scope control, etc.